+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-2.9 (unreleased)
+2.9 (2021-07-16)
 ----------------
 
 - Override `collective.iconifiedcategory` columns `category-column`,